Fix output of test_multithread_stress
The value printed as 'Total data:' is 0 if the test compressed less than 1 GB of
data. Convert tsize to double before division to make sure smaller values are
printed as well.

+++ test/test_multithread_stress.c
@@ -267,7 +267,7 @@ int main(int argc, char **argv)
 
 	printf("Testing finished...\n");
 
-	printf("Total data: %ld GB\n", tsize/(1<<30));
+	printf("Total data: %.3f GB\n", ((double)tsize)/(1<<30));
 
 	return failed_thread;
 }
